+++ tests/data_model/test_embedding_vectors.py
@@ -123,8 +123,8 @@ def test_vector1d_dm(float_seq_class, random_values, float_type):
         (dm.NpFloat64Sequence, random_numpy_vector1d_float64, np.float64),
     ],
 )
-def test_vector1d_dm_from_embeddings(float_seq_class, random_values, float_type):
-    v = dm.Vector1D.from_embeddings(random_values)
+def test_vector1d_dm_from_vector(float_seq_class, random_values, float_type):
+    v = dm.Vector1D.from_vector(random_values)
     assert isinstance(v.data, float_seq_class)
     assert isinstance(v.data.values[0], float_type)
     _assert_array_check(v, random_values, float_type)
